refactor: split attempts-helpers.ts into focused modules (#248, #252, #255)
Break up the 530-line god module into four single-responsibility files,
each with fewer external dependencies and independent testability.
No behavioral changes — all public exports maintain identical signatures.

Changes:
- Extract attempts-input.ts (~178 lines): sanitizeInput, toPromptHashPayload, getPdfProvenance, buildMetadata, stableSerialize
- Extract attempts-rate-limit.ts (~56 lines): selectUserGenerationAttemptWindowStats, computeRetryAfterSeconds
- Extract attempts-persistence.ts (~232 lines): persistSuccessfulAttempt, normalizeParsedModules, assertAttemptIdMatchesReservation, isAttemptsDbClient
- Reduce attempts-helpers.ts to ~70 lines: isProviderErrorRetryable, logAttemptEvent, getProviderErrorStatus
- Update consumer imports in attempts.ts, rate-limit.ts, orchestrator.ts
