+++ src/lib/db/queries/helpers/attempts-helpers.ts
@@ -1,67 +1,7 @@
 import { getCorrelationId } from '@/lib/api/context';
 import { appEnv } from '@/lib/config/env';
-import { hashSha256 } from '@/lib/crypto/hash';
-import { truncateToLength } from '@/lib/db/queries/helpers/truncation';
-import type {
-  AttemptError,
-  AttemptMetadata,
-  AttemptReservation,
-  AttemptsDbClient,
-  FinalizeSuccessPersistenceParams,
-  GenerationAttemptRecord,
-  MetadataParams,
-  NormalizedModuleData,
-  NormalizedModulesResult,
-  PdfProvenanceData,
-  SanitizedInput,
-  UserGenerationAttemptWindowStats,
-  UserGenerationAttemptsSinceParams,
-} from '@/lib/db/queries/types/attempts.types';
-import {
-  generationAttempts,
-  learningPlans,
-  modules,
-  tasks,
-} from '@/lib/db/schema';
-import { db as serviceDb } from '@/lib/db/service-role';
+import type { AttemptError } from '@/lib/db/queries/types/attempts.types';
 import { logger } from '@/lib/logging/logger';
-import {
-  aggregateNormalizationFlags,
-  normalizeModuleMinutes,
-  normalizeTaskMinutes,
-} from '@/shared/constants/effort';
-import { PLAN_GENERATION_WINDOW_MS } from '@/shared/constants/generation';
-import {
-  NOTES_MAX_LENGTH,
-  TOPIC_MAX_LENGTH,
-} from '@/shared/constants/learning-plans';
-import { and, count, eq, gte, min, sql, type SQL } from 'drizzle-orm';
-
-import type { ParsedModule } from '@/shared/types/ai-parser.types';
-import type { GenerationInput } from '@/shared/types/ai-provider.types';
-
-/** Drizzle-like methods required by attempt operations (reserve, finalize). */
-const ATTEMPTS_DB_METHODS = [
-  'select',
-  'insert',
-  'update',
-  'delete',
-  'transaction',
-] as const;
-
-/**
- * Type guard for AttemptsDbClient. Use when accepting db from unknown (e.g. options bags)
- * to fail fast with a clear error instead of obscure Drizzle errors later.
- */
-export function isAttemptsDbClient(db: unknown): db is AttemptsDbClient {
-  if (db == null || typeof db !== 'object') {
-    return false;
-  }
-  const obj = db as Record<string, unknown>;
-  return ATTEMPTS_DB_METHODS.every(
-    (method) => typeof obj[method] === 'function'
-  );
-}
 
 function getProviderErrorStatus(
   attemptErr: AttemptError | null | undefined
@@ -127,404 +67,4 @@ export function logAttemptEvent(
     // eslint-disable-next-line no-console
     console.info(`[attempts] ${event}`, enriched);
   }
-}
-
-export function stableSerialize(value: unknown): string {
-  if (value === undefined) {
-    return 'null';
-  }
-
-  if (value === null || typeof value !== 'object') {
-    return JSON.stringify(value);
-  }
-
-  if (Array.isArray(value)) {
-    return `[${value.map((entry) => stableSerialize(entry)).join(',')}]`;
-  }
-
-  const record = value as Record<string, unknown>;
-  const keys = Object.keys(record).toSorted();
-  return `{${keys.map((key) => `${JSON.stringify(key)}:${stableSerialize(record[key])}`).join(',')}}`;
-}
-
-function getPdfContextDigest(input: GenerationInput): string | null {
-  if (!input.pdfContext) {
-    return null;
-  }
-
-  return hashSha256(stableSerialize(input.pdfContext));
-}
-
-function hasPdfProvenanceInput(
-  input: GenerationInput
-): input is GenerationInput & {
-  pdfContext: NonNullable<GenerationInput['pdfContext']>;
-  pdfExtractionHash: string;
-  pdfProofVersion?: 1;
-} {
-  return (
-    input.pdfContext !== undefined &&
-    input.pdfContext !== null &&
-    typeof input.pdfExtractionHash === 'string' &&
-    input.pdfExtractionHash !== ''
-  );
-}
-
-export function getPdfProvenance(
-  input: GenerationInput
-): PdfProvenanceData | null {
-  if (!hasPdfProvenanceInput(input)) {
-    return null;
-  }
-
-  const contextDigest = getPdfContextDigest(input);
-  if (!contextDigest) {
-    return null;
-  }
-
-  return {
-    extractionHash: input.pdfExtractionHash,
-    proofVersion: input.pdfProofVersion ?? 1,
-    contextDigest,
-  };
-}
-
-export function buildMetadata(params: MetadataParams): AttemptMetadata {
-  const {
-    sanitized,
-    providerMetadata,
-    modulesClamped,
-    tasksClamped,
-    startedAt,
-    finishedAt,
-    extendedTimeout,
-    pdfProvenance,
-    failure,
-  } = params;
-
-  return {
-    input: {
-      topic: {
-        truncated: sanitized.topic.truncated,
-        original_length: sanitized.topic.originalLength,
-      },
-      notes:
-        sanitized.notes.originalLength !== undefined
-          ? {
-              truncated: sanitized.notes.truncated,
-              original_length: sanitized.notes.originalLength,
-            }
-          : null,
-    },
-    normalization: {
-      modules_clamped: modulesClamped,
-      tasks_clamped: tasksClamped,
-    },
-    timing: {
-      started_at: startedAt.toISOString(),
-      finished_at: finishedAt.toISOString(),
-      duration_ms: Math.max(
-        0,
-        Math.round(finishedAt.getTime() - startedAt.getTime())
-      ),
-      extended_timeout: extendedTimeout,
-    },
-    pdf: pdfProvenance
-      ? {
-          extraction_hash: pdfProvenance.extractionHash,
-          proof_version: pdfProvenance.proofVersion,
-          context_digest: pdfProvenance.contextDigest,
-        }
-      : null,
-    provider: providerMetadata ?? null,
-    failure: failure ?? null,
-  };
-}
-
-export function sanitizeInput(input: GenerationInput): SanitizedInput {
-  const topicResult = truncateToLength(input.topic, TOPIC_MAX_LENGTH);
-  if (topicResult.value === undefined) {
-    throw new Error('Topic is required for generation attempts.');
-  }
-
-  const topicValue = topicResult.value;
-
-  if (typeof topicValue !== 'string' || topicValue.trim().length === 0) {
-    throw new Error('GenerationInput.topic must be a non-empty string.');
-  }
-
-  const notesResult = truncateToLength(
-    input.notes ?? undefined,
-    NOTES_MAX_LENGTH
-  );
-
-  return {
-    topic: {
-      value: topicValue,
-      truncated: topicResult.truncated,
-      originalLength: topicResult.originalLength ?? topicValue.length,
-    },
-    notes: {
-      value: notesResult.value,
-      truncated: notesResult.truncated,
-      originalLength: notesResult.originalLength,
-    },
-  };
-}
-
-export function toPromptHashPayload(
-  planId: string,
-  userId: string,
-  input: GenerationInput,
-  sanitized: SanitizedInput
-): Record<string, unknown> {
-  const pdfContextDigest = getPdfContextDigest(input);
-
-  return {
-    planId,
-    userId,
-    topic: sanitized.topic.value,
-    notes: sanitized.notes.value ?? null,
-    skillLevel: input.skillLevel,
-    weeklyHours: input.weeklyHours,
-    learningStyle: input.learningStyle,
-    pdfExtractionHash: input.pdfExtractionHash ?? null,
-    pdfProofVersion: input.pdfProofVersion ?? null,
-    pdfContextDigest,
-  } satisfies Record<string, unknown>;
-}
-
-export function normalizeParsedModules(
-  modulesInput: ParsedModule[]
-): NormalizedModulesResult {
-  const moduleFlags = [] as ReturnType<typeof normalizeModuleMinutes>[];
-  const taskFlags = [] as ReturnType<typeof normalizeTaskMinutes>[];
-
-  const normalizedModules = modulesInput.map((module) => {
-    const normalizedModule = normalizeModuleMinutes(module.estimatedMinutes);
-    moduleFlags.push(normalizedModule);
-
-    const normalizedTasks = module.tasks.map((task) => {
-      const normalizedTask = normalizeTaskMinutes(task.estimatedMinutes);
-      taskFlags.push(normalizedTask);
-      return {
-        title: task.title,
-        description: task.description ?? null,
-        estimatedMinutes: normalizedTask.value,
-      };
-    });
-
-    return {
-      title: module.title,
-      description: module.description ?? null,
-      estimatedMinutes: normalizedModule.value,
-      tasks: normalizedTasks,
-    };
-  });
-
-  const normalizationFlags = aggregateNormalizationFlags(
-    moduleFlags,
-    taskFlags
-  );
-
-  return { normalizedModules, normalizationFlags };
-}
-
-function userAttemptsSincePredicate(
-  userId: string,
-  since: Date
-): SQL | undefined {
-  return and(
-    eq(learningPlans.userId, userId),
-    gte(generationAttempts.createdAt, since)
-  );
-}
-
-export async function selectUserGenerationAttemptWindowStats({
-  userId,
-  dbClient,
-  since,
-}: UserGenerationAttemptsSinceParams): Promise<UserGenerationAttemptWindowStats> {
-  const [row] = await dbClient
-    .select({
-      value: count(generationAttempts.id),
-      oldestCreatedAt: min(generationAttempts.createdAt),
-    })
-    .from(generationAttempts)
-    .innerJoin(learningPlans, eq(generationAttempts.planId, learningPlans.id))
-    .where(userAttemptsSincePredicate(userId, since));
-
-  return {
-    count: row?.value ?? 0,
-    oldestAttemptCreatedAt: row?.oldestCreatedAt ?? null,
-  };
-}
-
-export function computeRetryAfterSeconds(
-  oldestAttemptCreatedAt: Date | null,
-  now: Date
-): number {
-  if (!oldestAttemptCreatedAt) {
-    return Math.floor(PLAN_GENERATION_WINDOW_MS / 1000);
-  }
-
-  return Math.max(
-    0,
-    Math.floor(
-      (oldestAttemptCreatedAt.getTime() +
-        PLAN_GENERATION_WINDOW_MS -
-        now.getTime()) /
-        1000
-    )
-  );
-}
-
-export function assertAttemptIdMatchesReservation(
-  attemptId: string,
-  preparation: AttemptReservation
-): void {
-  if (attemptId !== preparation.attemptId) {
-    throw new Error('Attempt ID mismatch between params and reserved attempt.');
-  }
-}
-
-export async function persistSuccessfulAttempt(
-  params: FinalizeSuccessPersistenceParams
-): Promise<GenerationAttemptRecord> {
-  const {
-    attemptId,
-    planId,
-    preparation,
-    normalizedModules,
-    normalizationFlags,
-    modulesCount,
-    tasksCount,
-    durationMs,
-    metadata,
-    finishedAt,
-    dbClient,
-  } = params;
-
-  const shouldNormalizeRlsContext = dbClient !== serviceDb;
-  let requestJwtClaims: string | null = null;
-
-  if (shouldNormalizeRlsContext) {
-    const claimsRows = await dbClient.execute<{ claims: string | null }>(
-      sql`SELECT current_setting('request.jwt.claims', true) AS claims`
-    );
-    const rawClaims = claimsRows[0]?.claims;
-    if (typeof rawClaims === 'string' && rawClaims.length > 0) {
-      requestJwtClaims = rawClaims;
-    }
-  }
-
-  return dbClient.transaction(async (tx) => {
-    if (shouldNormalizeRlsContext && requestJwtClaims !== null) {
-      await tx.execute(
-        sql`SELECT set_config('request.jwt.claims', ${requestJwtClaims}, true)`
-      );
-    }
-
-    await tx.delete(modules).where(eq(modules.planId, planId));
-
-    const moduleValues = normalizedModules.map(
-      (normalizedModule: NormalizedModuleData, index: number) => ({
-        planId,
-        order: index + 1,
-        title: normalizedModule.title,
-        description: normalizedModule.description,
-        estimatedMinutes: normalizedModule.estimatedMinutes,
-      })
-    );
-    const insertedModuleRows =
-      moduleValues.length > 0
-        ? await tx
-            .insert(modules)
-            .values(moduleValues)
-            .returning({ id: modules.id })
-        : [];
-
-    if (insertedModuleRows.length !== normalizedModules.length) {
-      throw new Error('Failed to insert all modules for generation attempt.');
-    }
-
-    const taskValues: Array<{
-      moduleId: string;
-      order: number;
-      title: string;
-      description: string | null;
-      estimatedMinutes: number;
-    }> = [];
-
-    for (let i = 0; i < insertedModuleRows.length; i++) {
-      const moduleRow = insertedModuleRows[i];
-      const moduleEntry = normalizedModules[i];
-
-      if (!moduleRow || !moduleEntry) {
-        throw new Error('Failed to map inserted modules to generated tasks.');
-      }
-
-      for (
-        let taskIndex = 0;
-        taskIndex < moduleEntry.tasks.length;
-        taskIndex++
-      ) {
-        const task = moduleEntry.tasks[taskIndex];
-        if (!task) {
-          throw new Error('Failed to map generated task for insertion.');
-        }
-
-        taskValues.push({
-          moduleId: moduleRow.id,
-          order: taskIndex + 1,
-          title: task.title,
-          description: task.description,
-          estimatedMinutes: task.estimatedMinutes,
-        });
-      }
-    }
-
-    if (taskValues.length > 0) {
-      await tx.insert(tasks).values(taskValues);
-    }
-
-    const [attempt] = await tx
-      .update(generationAttempts)
-      .set({
-        status: 'success',
-        classification: null,
-        durationMs: Math.max(0, Math.round(durationMs)),
-        modulesCount,
-        tasksCount,
-        truncatedTopic: preparation.sanitized.topic.truncated,
-        truncatedNotes: preparation.sanitized.notes.truncated ?? false,
-        normalizedEffort:
-          normalizationFlags.modulesClamped || normalizationFlags.tasksClamped,
-        metadata,
-      })
-      .where(
-        and(
-          eq(generationAttempts.id, attemptId),
-          eq(generationAttempts.planId, planId),
-          eq(generationAttempts.status, 'in_progress')
-        )
-      )
-      .returning();
-
-    if (!attempt) {
-      throw new Error('Failed to finalize generation attempt as success.');
-    }
-
-    await tx
-      .update(learningPlans)
-      .set({
-        generationStatus: 'ready',
-        isQuotaEligible: true,
-        finalizedAt: finishedAt,
-        updatedAt: finishedAt,
-      })
-      .where(eq(learningPlans.id, planId));
-
-    return attempt;
-  });
 }
